+++ examples/src/main/kotlin/org/partiql/examples/EvaluationWithBindings.kt
@@ -1,6 +1,5 @@
 package org.partiql.examples
 
-import com.amazon.ion.system.IonSystemBuilder
 import org.partiql.examples.util.Example
 import org.partiql.lang.CompilerPipeline
 import org.partiql.lang.eval.Bindings
@@ -10,8 +9,7 @@ import java.io.PrintStream
 
 /** Demonstrates how to supply a global variable to the session. */
 class EvaluationWithBindings(out: PrintStream) : Example(out) {
-    val ion = IonSystemBuilder.standard().build()
-    val pipeline = CompilerPipeline.standard(ion)
+    val pipeline = CompilerPipeline.standard()
 
     override fun run() {
         // Compiles a simple expression containing a reference to a global variable.