+++ lang/src/main/kotlin/org/partiql/lang/eval/io/DelimitedValues.kt
@@ -22,12 +22,14 @@ import com.amazon.ion.IonSystem
 import com.amazon.ion.IonTimestamp
 import com.amazon.ion.IonType
 import com.amazon.ion.IonValue
+import com.amazon.ion.system.IonSystemBuilder
 import org.apache.commons.csv.CSVFormat
 import org.apache.commons.csv.CSVParser
 import org.apache.commons.csv.CSVPrinter
 import org.partiql.lang.eval.BindingCase
 import org.partiql.lang.eval.BindingName
 import org.partiql.lang.eval.ExprValue
+import org.partiql.lang.eval.ExprValueFactory
 import org.partiql.lang.eval.StructOrdering
 import org.partiql.lang.eval.namedValue
 import org.partiql.lang.eval.orderedNames
@@ -49,9 +51,21 @@ object DelimitedValues {
     enum class ConversionMode {
         /** Attempt to parse each value as a scalar, and fall back to string. */
         AUTO {
-            override fun convert(ion: IonSystem, raw: String): ExprValue = try {
-                val converted = ion.singleValue(raw)
+            override fun convert(valueFactory: ExprValueFactory, raw: String): ExprValue = try {
+                val converted = valueFactory.ion.singleValue(raw)
                 when (converted) {
+                    is IonInt, is IonFloat, is IonDecimal, is IonTimestamp ->
+                        valueFactory.newFromIonValue(converted)
+                    // if we can't convert the above, we just use the input string as-is
+                    else -> valueFactory.newString(raw)
+                }
+            } catch (e: IonException) {
+                valueFactory.newString(raw)
+            }
+
+            override fun convert(raw: String): ExprValue = try {
+                val ion = IonSystemBuilder.standard().build()
+                when (val converted = ion.singleValue(raw)) {
                     is IonInt, is IonFloat, is IonDecimal, is IonTimestamp ->
                         ExprValue.of(converted)
                     // if we can't convert the above, we just use the input string as-is
@@ -63,10 +77,15 @@ object DelimitedValues {
         },
         /** Each field is a string. */
         NONE {
-            override fun convert(ion: IonSystem, raw: String): ExprValue = ExprValue.newString(raw)
+            override fun convert(valueFactory: ExprValueFactory, raw: String): ExprValue = valueFactory.newString(raw)
+
+            override fun convert(raw: String): ExprValue = ExprValue.newString(raw)
         };
 
-        abstract fun convert(ion: IonSystem, raw: String): ExprValue
+        @Deprecated("[ExprValueFactory] is deprecated. Please use `convert(row: String): ExprValue` instead")
+        abstract fun convert(valueFactory: ExprValueFactory, raw: String): ExprValue
+
+        abstract fun convert(raw: String): ExprValue
     }
 
     /**
@@ -95,7 +114,7 @@ object DelimitedValues {
                         i < columns.size -> columns[i]
                         else -> syntheticColumnName(i)
                     }
-                    conversionMode.convert(ion, value).namedValue(ExprValue.newString(name))
+                    conversionMode.convert(value).namedValue(ExprValue.newString(name))
                 },
                 StructOrdering.ORDERED
             )